requirements.txt
Flask
flask-pymongo
PyMongo
tensorflow
Pillow
werkzeug
flask-mail